info components moved from css modules to emotion css (#390)
* info components moved from css modules to emotion css

* fixed import and deleted css module

* prettier format
