--- app/components/views/Info/AddToMetaMaskButton/index.module.css
@@ -1,8 +1,0 @@
-.addToMetaMaskButton {
-  padding: 0.8rem;
-  margin-inline-start: -0.8rem;
-}
-.addToMetaMaskButton img {
-  width: 1.6rem;
-  height: 1.6rem;
-}
