+++ app/components/views/Info/AddToMetaMaskButton/styles.ts
@@ -1,0 +1,10 @@
+import { css } from "@emotion/css";
+
+export const addToMetaMaskButton = css`
+  padding: 0.8rem;
+  margin-inline-start: -0.8rem;
+  img {
+    width: 1.6rem;
+    height: 1.6rem;
+  }
+`;
