+++ app/components/views/Info/CopyAddressButton/styles.ts
@@ -1,0 +1,17 @@
+import { css } from "@emotion/css";
+
+// used !important here bc .Home-container class is more specific than this one
+export const copyAddressButtonIcon = css`
+  height: 1.6rem !important;
+  width: 1.6rem !important;
+  cursor: pointer;
+`;
+
+export const checkmark = css`
+  padding: 0.8rem;
+  max-height: 3.2rem;
+`;
+
+export const copyAddressButton = css`
+  padding: 0.8rem;
+`;
